+++ temporalio/worker/_interceptor.py
@@ -174,6 +174,7 @@ class ContinueAsNewInput:
     headers: Mapping[str, temporalio.api.common.v1.Payload]
     versioning_intent: VersioningIntent | None
     initial_versioning_behavior: ContinueAsNewVersioningBehavior | None
+    event_groups: Sequence[temporalio.workflow.EventGroup] | None
     # The types may be absent
     arg_types: list[type] | None
 
@@ -260,6 +261,7 @@ class StartActivityInput:
     disable_eager_execution: bool
     versioning_intent: VersioningIntent | None
     summary: str | None
+    event_groups: Sequence[temporalio.workflow.EventGroup] | None
     priority: temporalio.common.Priority
     # The types may be absent
     arg_types: list[type] | None
@@ -290,6 +292,7 @@ class StartChildWorkflowInput:
     versioning_intent: VersioningIntent | None
     static_summary: str | None
     static_details: str | None
+    event_groups: Sequence[temporalio.workflow.EventGroup] | None
     priority: temporalio.common.Priority
     # The types may be absent
     arg_types: list[type] | None
@@ -310,6 +313,7 @@ class StartNexusOperationInput(Generic[InputT, OutputT]):
     cancellation_type: temporalio.workflow.NexusOperationCancellationType
     headers: Mapping[str, str] | None
     summary: str | None
+    event_groups: Sequence[temporalio.workflow.EventGroup] | None = None
     output_type: type[OutputT] | None = None
 
     def __post_init__(self) -> None:
@@ -363,6 +367,7 @@ class StartLocalActivityInput:
     cancellation_type: temporalio.workflow.ActivityCancellationType
     headers: Mapping[str, temporalio.api.common.v1.Payload]
     summary: str | None
+    event_groups: Sequence[temporalio.workflow.EventGroup] | None
 
     # The types may be absent
     arg_types: list[type] | None